# docs/openapi/admin-api.yaml
openapi: "3.0.3"
info:
  title: "PJB Admin API"
  version: "1.0.0-frontend-freeze"
  description: "Contrato exportado estaticamente para integração com frontend."
servers:
  -
    url: "http://localhost:8080"
paths:
...
  /api/admin/ritos/diagnostico/processo/{processoId}:
    get:
      summary: "AdminRitoDiagnosticsController::GET /api/admin/ritos/diagnostico/processo/{processoId}"
      tags:
        - "com.tcc.pjb.backend.controller.admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/quality-gates/architecture:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/architecture"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/blockers:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/blockers"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/build:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/build"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/quality-gates/contracts:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/contracts"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/dast:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/dast"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/integration:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/integration"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/matrix:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/matrix"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/mutation:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/mutation"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
  /api/v1/admin/quality-gates/summary:
    get:
      summary: "AdminQualityGatesController::GET /api/v1/admin/quality-gates/summary"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/trabalhista/processos/{processoId}:
    get:
      summary: "AdminTrabalhistaController::GET /api/v1/admin/trabalhista/processos/{processoId}"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
  /api/v1/admin/trabalhista/processos/{processoId}/execucao:
    get:
      summary: "AdminTrabalhistaController::GET /api/v1/admin/trabalhista/processos/{processoId}/execucao"
      tags:
        - "admin"
      responses:
        200:
          description: "OK"
      security:
        -
          bearerAuth:
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
    basicAuth:
      type: "http"
      scheme: "basic"